requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.6.0
    # via django
backports-zoneinfo==0.2.1
    # via django
django==4.1.4
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.13.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
mysqlclient==2.1.1
    # via -r requirements.in
pyjwt==2.6.0
    # via djangorestframework-simplejwt
pytz==2022.7
    # via djangorestframework
sqlparse==0.4.3
    # via django
